+++ CHANGELOG.md
@@ -21,6 +21,7 @@
     - maven-surefire 2.22.1
     - h2 1.4.198
 - Fix workflow history cleanup to keep the actions that hold the latest values of state variables
+- nFlow Explorer: Custom content to workflow definition and workflow instance pages. 
 
 ## 5.3.3 (2019-02-04)